Rename feature directory to features
