+++ turing-calls/src/App.tsx
@@ -3,12 +3,12 @@ import { useEffect } from 'react'
 import { Navigate, Route, Routes, useNavigate } from 'react-router-dom'
 
 import './App.css'
-import { Login } from './feature/login/Login.tsx'
-import { Calls } from './feature/calls/Calls.tsx'
+import { Login } from './features/login/Login.tsx'
+import { Calls } from './features/calls/Calls.tsx'
 import { AppHeader } from './components/AppHeader.tsx'
 import { AuthRoute } from './components/AuthRoute.tsx'
-import { useLogin } from './feature/login/useLogin.ts'
-import { User } from './feature/login/types.ts'
+import { useLogin } from './features/login/useLogin.ts'
+import { User } from './features/login/types.ts'
 
 const { Header, Content } = Layout
