+++ turing-calls/src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
-import { LoginProvider } from './feature/login/LoginProvider.tsx'
+import { LoginProvider } from './features/login/LoginProvider.tsx'
 import { BrowserRouter } from 'react-router-dom'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
